+++ CHANGES.md
@@ -18,6 +18,8 @@ Version 1.5
 * Support for standalone-deriving and default signatures in GHC 7.10.
   This means that there are now two new constructors for `DDec`.
 
+* Support for `static` expressions, which are new in GHC 7.10.
+
 Version 1.4.2
 -------------
 * `expand` functions now consider open type families, as long as the type